+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Add the `permissions` option for setting the file mode for the `unix-socket` source
 - Tests can be run without their suite. [#1238](https://github.com/tremor-rs/tremor-runtime/pull/1283)
 - Add the `std::size` module to convert sizes
+- Add custom function calls to constant folding
 
 ### Fixes